Move banner printing function to dedicated startup utils file
