+++ growth_stock_screener/run_screen.py
@@ -13,8 +13,8 @@ print(
 # run screen iterations
 import iterations.nasdaq_listings
 import iterations.relative_strength
+import iterations.liquidity
 
-# import iterations.liquidity
 # import iterations.trend
 # import iterations.revenue_growth
 # import iterations.institutional_accumulation
